--- app/src/test/java/samples/TestController.java
@@ -1,22 +1,0 @@
-package samples;
-
-import com.interface21.context.stereotype.Controller;
-import com.interface21.web.bind.annotation.RequestMapping;
-import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class TestController {
-
-    private static final Logger log = LoggerFactory.getLogger(TestController.class);
-
-    @RequestMapping(value = "/get-test", method = RequestMethod.GET)
-    public ModelAndView test(final HttpServletRequest request, final HttpServletResponse response) {
-        return new ModelAndView(new JspView("/test.jsp"));
-    }
-}
