+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerMappingRegistry.java
@@ -1,4 +1,4 @@
-package camp.nextstep;
+package com.interface21.webmvc.servlet;
 
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
