+++ mvc/src/test/java/samples/TestController.java
@@ -1,19 +1,29 @@
 package samples;
 
 import com.interface21.context.stereotype.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.view.JspView;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @Controller
 public class TestController {
 
     private static final Logger log = LoggerFactory.getLogger(TestController.class);
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView root(final HttpServletRequest request, final HttpServletResponse response) {
+        return ModelAndView.createJspView("/index.jsp");
+    }
+
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    public ModelAndView registerView(final HttpServletRequest request, final HttpServletResponse response) {
+        return ModelAndView.createJspView("/register.jsp");
+    }
 
     @RequestMapping(value = "/get-test", method = RequestMethod.GET)
     public ModelAndView findUserId(final HttpServletRequest request, final HttpServletResponse response) {
